olive/requirements.txt
Flask==3.0.0
flask-login==0.6.3
PyYAML==6.0.1
Authlib==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
